+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Economy service, `get_player_customization` method, and associated response model.
+- Exposed token updating for `HaloInfiniteClient` via a `set_tokens` method.
 
 ### Changed